Bug fixes for received messages having no name next to them and the app icon not appearing.
Flattened the app icon.

Implemented skeleton code for the B to A key transfer. This is messy (because it has to run as a scheduled task because NFC only triggers an event when it comes into range or when a beam is received, not when a beam is sent, so on the sending device I can't tell whether a beam has been sent or not), and I don't like it, but we'll see what Shujun says. This code is currently commented so that it's non-functional, so there's warnings about unused code in this commit.